Friend APIs, FriendUI, EventAPI Changes, JWT Changes
Edited Event APIs to return what they do on main (not just the token) and made friends into an array of strings.
Changed Calendar UI to load friends into a form, so you can select multiple when adding event.
Added Friend CRUD APIs and a UI to test.
Added HandleTokenExpiration which checks if the error is due to JWT expired, gives an alert and redirects to login.